fix password login
