+++ src/routes/Landing.tsx
@@ -1,5 +1,5 @@
-import { Box, Button, Card, Image, Modal, Text, TextInput, useMediaQuery } from '@0xsequence/design-system'
-import { ChangeEvent, useEffect, useState } from 'react'
+import { Box, Button, Card, Image, Text, useMediaQuery } from '@0xsequence/design-system'
+import { useEffect, useState } from 'react'
 import { Link } from 'react-router-dom'
 
 import { useObservable, useStore } from '~/stores'
@@ -10,6 +10,7 @@ import walletIcon from '~/assets/icons/wallet.svg'
 import bgImageMobile from '~/assets/images/recovery-wallet-bg-mobile.jpg'
 import bgImage from '~/assets/images/recovery-wallet-bg.jpg'
 import SequenceRecoveryLogo from '~/assets/images/sequence-wallet-recovery.svg'
+import { PasswordUnlock } from '~/components/auth/PasswordUnlock'
 
 const desktopBg = {
   backgroundImage: `url(${bgImage})`,
@@ -19,37 +20,14 @@ const desktopBg = {
 
 export default function Landing() {
   const isMobile = useMediaQuery('isMobile')
-
   const authStore = useStore(AuthStore)
   const isLoadingAccountObservable = useObservable(authStore.isLoadingAccount)
 
-  const [password, setPassword] = useState('')
-  const [isReseting, setIsReseting] = useState(false)
-  const [wrongPassword, setWrongPassword] = useState(false)
   const [isLoadingAccount, setIsLoadingAccount] = useState(false)
 
   useEffect(() => {
     setIsLoadingAccount(isLoadingAccountObservable)
   }, [isLoadingAccountObservable])
-
-  const handleUnlock = async () => {
-    try {
-      await authStore.loadAccount(password)
-    } catch (e) {
-      console.warn(e)
-      setWrongPassword(true)
-    }
-  }
-
-  const handleResetConfirmation = () => {
-    setIsReseting(true)
-  }
-
-  const handleReset = () => {
-    authStore.logout()
-    setIsReseting(false)
-    authStore.isLoadingAccount.set(false)
-  }
 
   return (
     <Box
@@ -90,53 +68,7 @@ export default function Landing() {
         </Box>
 
         {isLoadingAccount ? (
-          <>
-            <Text
-              variant="normal"
-              fontWeight="medium"
-              textAlign="center"
-              color="text80"
-              paddingX={isMobile ? '8' : undefined}
-              style={{ marginBottom: '-16px' }}
-            >
-              Enter your password to continue and unlock your wallet
-            </Text>
-            <Box flexDirection="column" gap="4" width="full">
-              <Box flexDirection="column" gap="1">
-                <Text variant="normal" fontWeight="medium" color="text80">
-                  Password
-                </Text>
-                <TextInput
-                  type="password"
-                  value={password}
-                  autoFocus
-                  onKeyPress={(ev: KeyboardEvent) => {
-                    if (ev.key === 'Enter') {
-                      handleUnlock()
-                    }
-                  }}
-                  onChange={(ev: ChangeEvent<HTMLInputElement>) => {
-                    setPassword(ev.target.value)
-                    setWrongPassword(false)
-                  }}
-                />
-                {wrongPassword && (
-                  <Text variant="small" color="negative" marginLeft="2" marginTop="1">
-                    Incorrect password
-                  </Text>
-                )}
-              </Box>
-              <Box flexDirection="row" justifyContent="flex-end" gap="4">
-                <Button
-                  label="Forgot password?"
-                  variant="text"
-                  shape="square"
-                  onClick={() => handleResetConfirmation()}
-                />
-                <Button label="Continue" variant="primary" shape="square" onClick={() => handleUnlock()} />
-              </Box>
-            </Box>
-          </>
+          <PasswordUnlock redirectOnSuccess={true} />
         ) : (
           <>
             <Box gap="2">
@@ -175,24 +107,6 @@ export default function Landing() {
           </>
         )}
       </Box>
-
-      {isReseting && (
-        <Modal size="md" onClose={() => setIsReseting(false)}>
-          <Box flexDirection="column" padding="6" gap="6">
-            <Text variant="large" color="text100" marginRight="8">
-              Are you sure you want to sign out?
-            </Text>
-            <Text variant="normal" fontWeight="medium" color="text50">
-              If you do not remember your password, you can reset and start over.
-              <br /> This will require you to re-enter your mnemonic.
-            </Text>
-            <Box flexDirection="row" justifyContent="flex-end" gap="2">
-              <Button label="Yes, reset" shape="square" variant="primary" onClick={() => handleReset()} />
-              <Button label="Cancel" shape="square" onClick={() => setIsReseting(false)} />
-            </Box>
-          </Box>
-        </Modal>
-      )}
     </Box>
   )
 }
